Curl support
Not all builds will be made from nixOS or a system with the nix package manager. This includes builds that are new installs and deployments that do not contain the git package. Consequently, these modules must be curled for installation. So far, I do not own a private and sufficiently secure domain. In the meantime, I will curl each module to https://0x0.st and will list the links for each.