Added i3wm and Polybar via Home-manager
It looks that xmonad would have been a nice touch to my nixOS setup but
that would remove the configurability that nix adds. Having a window
manager in haskell would facilitate a decent environment to program in
but it would remove portability via deployment. Also, much of the great
environment provided by haskell would be removed as it would be wrapped
in nix lang. Things such as djinn, ghci, and stack repl would be unable
to be used with this wrapper. For xmonad to be a for seeable future, it
must be individually included in a separate file. No nix configurations.
Just simply writing a haskell configuration for `xmonad.hs`. In the
meantime, i3 can be fully writen in nix lang as well as polybar. SO I
went forth to port all of my i3 config and wrote a polybar config to
match in nix. All of this is fully portable via home-manager. The
home-manager package is imported in the configuration.nix. For more
details, install home-manager and run `man home-configuration.nix`.