Changed Mpd to home.nix
It is rightfully so that the music player daemon shall be governed and
ran by each individual user instead of the root. Each user deserves its
own music library and unique daemon hosted to the loopback. As so, a
document to record the open ports of the loopback shall be made later.