Adding Install Script
Note: This will not really install but it will set up everything for
installation. This means you must still manually invoke the
`nixos-install` command at the shell. This is not included because the
hardware module and hardware patches must be commented out in the
`configuration.nix`. To do this automatically, the script would have to
comment those lines. And I'm not going to upload a version without those
lines uncommented because any installed build could never `nix-build`
without manual labor of uncommenting those lines. When looking at the
necessary requirements to automate this, it becomes very obvious on why
this design decision was made instead of the more short term convenient
way.

PS: there are plans for this automation along with hardware detection,
multi drive support, etc.